refactor and minor changes based on review
