+++ app/controllers/dashboard/access_tokens_controller.rb
@@ -1,5 +1,5 @@
 class Dashboard::AccessTokensController < Dashboard::BaseController
   def index
-    @access_tokens = current_account.access_tokens.page(params[:page]).per(10)
+    @access_tokens = current_account.access_tokens
   end
 end
