+++ app/controllers/dashboard/users_controller.rb
@@ -1,5 +1,5 @@
 class Dashboard::UsersController < Dashboard::BaseController
   def index
-    @users = current_account.users.page(params[:page]).per(10)
+    @users = current_account.users.page(params[:page])
   end
 end
