+++ app/views/dashboard/access_tokens/index.html.haml
@@ -14,5 +14,3 @@
           %td= index + 1
           %td= access_token.token
           %td= access_token.created_at
-.col-12
-  = paginate @access_tokens
