+++ app/views/dashboard/callouts/_form.html.haml
@@ -1,4 +1,4 @@
-= simple_form_for callout, url: url do |f|
+= simple_form_for [:dashboard, callout] do |f|
   = f.hidden_field :id
   .card
     .card-body
