+++ app/views/dashboard/callouts/edit.html.haml
@@ -1,3 +1,3 @@
 .col-12.col-md-8
   %h5.card-title Edit callout
-  = render partial: 'form', locals: { callout: @callout, url: dashboard_callout_path(@callout) }
+  = render partial: 'form', locals: { callout: @callout }
