+++ app/views/dashboard/callouts/new.html.haml
@@ -1,3 +1,3 @@
 .col-12.col-md-8
   %h5.card-title New callout
-  = render partial: 'form', locals: { callout: @callout, url: dashboard_callouts_path }
+  = render partial: 'form', locals: { callout: @callout }
